EVG-13370: move user lifecycle code into user package and cleanup user package methods
